+++ app/src/main/java/ai/heart/classickbeats/data/record/cache/TimelineRemoteKeyDao.kt
@@ -12,7 +12,7 @@ interface TimelineRemoteKeyDao {
     suspend fun insertAll(remoteKey: List<TimelineRemoteKey>)
 
     @Query("SELECT * FROM timeline_remote_keys WHERE timelineId = :timelineId")
-    suspend fun remoteKeysTimelineId(timelineId: Int): TimelineRemoteKey?
+    suspend fun remoteKeysTimelineId(timelineId: Long): TimelineRemoteKey?
 
     @Query("DELETE FROM timeline_remote_keys")
     suspend fun clearRemoteKeys()
